# ARTIQ Pull Request

## Description of Changes
RPC performance optimization for Python code. This patch reduces the number of calls to socket functions, thus hope to improve the throughput for small but more complex RPCs.

This patch also added some tests for small RPC calls, and print a table for performance test summary.

### Benchmark
Each test is run for 10 times. H2D means host to device, D2H means device to host.

artiq-zynq:

| Test                 | Mean before (MiB/s) |  std before (MiB/s) | Mean after (MiB/s) |  std after (MiB/s) | Mean Diff  |
| -------------------- | ------------: | ------------: | ------------: | ------------: | -------: |
| I32 Array (1MB) H2D  |        40.77 |         2.73 |        38.51 |         3.13 | -5.54% |
| I32 Array (1MB) D2H  |        28.77 |         2.51 |        32.24 |         0.65 | 12.06% |
| I32 Array (1KB) H2D  |         2.48 |         0.34 |         3.67 |         0.73 | 47.98% |
| I32 Array (1KB) D2H  |         2.20 |         0.24 |         3.15 |         0.29 | 43.18% |
| Bytes List (1MB) H2D |        35.78 |         2.73 |        32.76 |         3.67 | -8.44% |
| Bytes List (1MB) D2H |        31.57 |         1.75 |        29.84 |         2.97 | -5.48% |
| Bytes List (1KB) H2D |         2.53 |         0.30 |         3.66 |         0.59 | 44.66% |
| Bytes List (1KB) D2H |         2.39 |         0.08 |         3.34 |         0.38 | 39.75% |
| Bytes (1MB) H2D      |        53.80 |         3.36 |        52.78 |         3.16 | -1.90% |
| Bytes (1MB) D2H      |        41.10 |         1.90 |        43.47 |         0.37 |  5.77% |
| Bytes (1KB) H2D      |         2.38 |         0.29 |         3.83 |         0.69 | 60.92% |
| Bytes (1KB) D2H      |         2.27 |         0.25 |         3.44 |         0.30 | 51.54% |
| I32 List (1MB) H2D   |        21.78 |         1.77 |        31.58 |         4.89 | 45.00% |
| I32 List (1MB) D2H   |        27.35 |         0.35 |        25.71 |         2.95 | -6.00% |
| I32 List (1KB) H2D   |         2.51 |         0.33 |         3.67 |         0.70 | 46.22% |
| I32 List (1KB) D2H   |         2.34 |         0.27 |         3.39 |         0.48 | 44.87% |



KC705

| Test                 | Mean before (MiB/s) |  std before (MiB/s) | Mean after (MiB/s) |  std after (MiB/s) | Mean Diff  |
| -------------------- | ------------: | ------------: | ------------: | ------------: | -------: |
| I32 Array (1MB) H2D  |         2.06 |         0.00 |        2.05 |         0.01 | -0.49%|
| I32 Array (1MB) D2H  |         0.42 |         0.00 |        0.43 |         0.00 | +2.38%|
| I32 Array (1KB) H2D  |         0.29 |         0.00 |        0.34 |         0.02 | +17.24%|
| I32 Array (1KB) D2H  |         0.19 |         0.00 |        0.22 |         0.01 | +15.79%|
| Bytes List (1MB) H2D |         2.01 |         0.03 |        2.01 |         0.02 | +0.00%|
| Bytes List (1MB) D2H |         2.14 |         0.02 |        2.12 |         0.05 | -0.93%|
| Bytes List (1KB) H2D |         0.29 |         0.01 |        0.35 |         0.01 | +20.69%|
| Bytes List (1KB) D2H |         0.30 |         0.00 |        0.39 |         0.01 | +30.00%|
| Bytes (1MB) H2D      |         2.06 |         0.00 |        2.07 |         0.00 | +0.49%|
| Bytes (1MB) D2H      |         2.23 |         0.00 |        2.23 |         0.00 | +0.00%|
| Bytes (1KB) H2D      |         0.29 |         0.01 |        0.36 |         0.00 | +24.14%|
| Bytes (1KB) D2H      |         0.32 |         0.01 |        0.40 |         0.00 | +25.00%|
| I32 List (1MB) H2D   |         1.96 |         0.04 |        2.03 |         0.03 | +3.57%|
| I33 List (1MB) D2H   |         0.43 |         0.00 |        0.42 |         0.00 | -2.33%|
| I32 List (1KB) H2D   |         0.29 |         0.01 |        0.37 |         0.02 | +27.59%|
| I32 List (1KB) D2H   |         0.20 |         0.00 |        0.23 |         0.01 | +15.00%|



## Type of Changes

<!-- Leave ONLY the corresponding lines for the applicable type of change: -->
|   | Type |
| ------------- | ------------- |
| ✓  | :hammer: Refactoring  |

## Steps (Choose relevant, delete irrelevant before submitting)

### All Pull Requests

- [x] Use correct spelling and grammar.

### Code Changes

- [x] Run `flake8` to check code style (follow PEP-8 style). `flake8` has issues with parsing Migen/gateware code, ignore as necessary.
- [x] Test your changes or have someone test them. Mention what was tested and how.
- [x] Check, test, and update the [unittests in /artiq/test/](../artiq/test/) or [gateware simulations in /artiq/gateware/test](../artiq/gateware/test)

### Git Logistics

- [x] Split your contribution into logically separate changes (`git rebase --interactive`). Merge/squash/fixup commits that just fix or amend previous commits. Remove unintended changes & cleanup. See [tutorial](https://www.atlassian.com/git/tutorials/rewriting-history/git-rebase).
- [x] Write short & meaningful commit messages. Review each commit for messages (`git show`). Format:
  ```
  topic: description. < 50 characters total.
  
  Longer description. < 70 characters per line
  ```

### Licensing

See [copyright & licensing for more info](https://github.com/m-labs/artiq/blob/master/CONTRIBUTING.rst#copyright-and-sign-off).
ARTIQ files that do not contain a license header are copyrighted by M-Labs Limited and are licensed under LGPLv3+.